+++ NEWS.md
@@ -6,6 +6,7 @@
 * Added `UrlParameters` argument to the `Survey` class to specify what parameters should be saved in the database.
 * Added `pauseBuild` arg to `SurveyModel.build()` to prevent the survey from being built when created.
 * Added `surveyFromJson()` function to create a survey from a json file made with the visual creator.
+* Themes have been implemented. See `themeFile` argument in the `SurveyModel` class.
 
 ## 0.3.2